Adjust for revocation of microsoft-windows-2011
The certificate named "Microsoft Windows Production PCA 2011", bundled
in this repository as "microsoft-windows-2011.crt", was revoked by
Microsoft due to the BlackLotus vulnerability.

Add a note about it in extracerts/README.md, and move it to a different
directory, named "extracerts.REVOKED/". This ensures that users can
still access it if they need it, and are informed in the README about
(a) why they shouldn't use it, as well as (b) possible issues they may
face by not using it (primarily the inability to boot older Windows).