Provide support for prebuilt efitools
Some distributions are dropping support for efitools in their repos.
Somewhat understandable, as the software is essentially abandonware at
this point. However it's required by the scripts in this repository.
So a prebuilt binary package can now be used instead of the distro's
package, at least on x86_64. See 'efitools/README.md' for details.