Implement proper handling of owner GUIDs
Firstly, handle self-UUID-generation more robustly. Determine the best
way to generate a UUID (which may not be uuidgen), and do not allow
certain scripts to continue if no suitable UUID generation method was
found.

Automatically generate a random GUID with generate-keys.sh, which will
be the owner GUID for the user's own key set.

Support .crt.guid files under extracerts, which allow specifying a
hardcoded GUID for specific certificates (e.g., Microsoft).

Provide graceful fallbacks if hardcoded GUIDs are not found for some
extracert certificates, and same if the user's own key set does not
have an owner GUID file (which may be the case if it was generated from
an older commit on this repository). Handle this by just letting
create-databases.sh randomly generate any undefined GUIDs at runtime.